"Refactor user search method to search by username explicitly"
Updated the respective methods which are searching user to allow user search explicitly by their username instead of using a only single method to retrieve the user.

This was done so that there is clear differentiation in our codebase for searching a user by their username as it may look for a match in either the user's ID or their username depending on perspective. This way, it should be easier for future developers to understand the code's intention and reduce potential confusion.

Furthermore, adjusted the respective tests that were dependent on the old function name. Also, extra functions have been added to the UserController to handle the new methods for retrieving user.

IUserRepository interface, TaskEngine, UserInMemoryRepository, UserRepository classes were updated to reflect the new method. Tests related to User creation were also updated accordingly. These changes are reflected in the UsersController as well, where the respective endpoints are added and updated.