Change method parameters from string id to Guid id
The parameters for getting user tasks in TaskRepository, TaskEngine, ITaskEngine, and ITaskRepository have been changed from string id to Guid id to maintain consistency across the project. This will reduce confusion, and chances of error due to data type mismatches. Furthermore, the ternary operator in TaskInMemoryRepository's GetUserTasksByNameAsync and GetTasks methods was removed to simplify the retrieval of tasks, regardless if the user or task list is empty.