Refactor GetUserByName endpoint for better UX
Updated the GetUserByName action method in UsersController to accept a query parameter instead of a path parameter. This modification was implemented to make searching for users by name more intuitive for the users of the API. Also, to comply with REST principles, the endpoint name has been changed to 'search' implying a search operation. Each returned user now gets wrapped in an array to facilitate frontend handling of the response.