Add scripts for user creation in TaskManagement API
Two PowerShell scripts are added in 'Cnvs.Demo.TaskManagement.WebApi.Curl' for creating users in the TaskManagement API. CreateOneUser.ps1 creates a single user, while CreateThreeUsers.ps1 creates three users. The scripts use 'Invoke-RestMethod' cmdlet for making POST requests to the API with user details in the body. This will automate and ease the process of creating users for