Refactor GetUserTasks method in UsersController
Changed the variable name 'userAsync' to 'user' for better readability within the GetUserTasks method. Also used 'is' keyword for checking if 'user' is a NullUser instead of the Equals method for cleaner code. Lastly, updated the BadRequest response to NotFound for a missing user to better align with HTTP response conventions.