Add user assignment and history to Task entity
Added fields to track the user assigned to a task and the history of its assignments. This was done to facilitate managing task assignments and understanding the task's history. Now, it's possible to know how many times a task was transferred and who were the assigned users in the past.